+++ library/src/actions/parseBoolean/parseBoolean.test.ts
@@ -1,95 +1,95 @@
 import { describe, expect, test } from 'vitest';
 import {
-  stringbool,
-  type StringboolAction,
-  type StringboolIssue,
-  type StringboolOptions,
-} from './stringbool.ts';
+  parseBoolean,
+  type ParseBooleanAction,
+  type ParseBooleanConfig,
+  type ParseBooleanIssue,
+} from './parseBoolean.ts';
 
-describe('stringbool', () => {
+describe('parseBoolean', () => {
   const defaultExpected =
-    '("true" | "1" | "yes" | "y" | "on" | "enabled" | "false" | "0" | "no" | "n" | "off" | "disabled")';
+    '(true | 1 | "true" | "1" | "yes" | "y" | "on" | "enabled" | false | 0 | "false" | "0" | "no" | "n" | "off" | "disabled")';
 
   describe('should return action object', () => {
-    test('for default options', () => {
-      expect(stringbool()).toStrictEqual({
-        kind: 'transformation',
-        type: 'stringbool',
-        reference: stringbool,
+    const config: ParseBooleanConfig = {
+      truthy: ['yep'],
+      falsy: ['nope'],
+    };
+    const baseAction: Omit<
+      ParseBooleanAction<unknown, never, never>,
+      'message' | 'config' | 'expects'
+    > = {
+      kind: 'transformation',
+      type: 'parse_boolean',
+      reference: parseBoolean,
+      async: false,
+      '~run': expect.any(Function),
+    };
+
+    test('with undefined config and undefined message', () => {
+      const action: ParseBooleanAction<unknown, undefined, undefined> = {
+        ...baseAction,
         expects: defaultExpected,
-        options: {
-          truthy: ['true', '1', 'yes', 'y', 'on', 'enabled'],
-          falsy: ['false', '0', 'no', 'n', 'off', 'disabled'],
-          case: 'insensitive',
-        },
-        async: false,
-        '~run': expect.any(Function),
-      } satisfies StringboolAction<unknown>);
+        config: undefined,
+        message: undefined,
+      };
+      expect(parseBoolean()).toStrictEqual(action);
+      expect(parseBoolean(undefined)).toStrictEqual(action);
+      expect(parseBoolean(undefined, undefined)).toStrictEqual(action);
     });
 
-    test('for custom options truthy only', () => {
-      const options: StringboolOptions = {
-        truthy: ['yep'],
-      };
-
-      expect(stringbool(options)).toStrictEqual({
-        kind: 'transformation',
-        type: 'stringbool',
-        reference: stringbool,
-        expects: '("yep" | "false" | "0" | "no" | "n" | "off" | "disabled")',
-        options: {
-          truthy: ['yep'],
-          falsy: ['false', '0', 'no', 'n', 'off', 'disabled'],
-          case: 'insensitive',
-        },
-        async: false,
-        '~run': expect.any(Function),
-      } satisfies StringboolAction<unknown>);
+    test('with undefined config and string message', () => {
+      expect(parseBoolean(undefined, 'message')).toStrictEqual({
+        ...baseAction,
+        expects: defaultExpected,
+        config: undefined,
+        message: 'message',
+      } satisfies ParseBooleanAction<unknown, undefined, 'message'>);
     });
 
-    test('for custom options case undefined', () => {
-      const options: StringboolOptions = {
-        truthy: ['yep'],
-        falsy: ['nope'],
-        case: undefined,
-      };
+    test('with undefined config and function message', () => {
+      const message = () => 'message';
+      expect(parseBoolean(undefined, message)).toStrictEqual({
+        ...baseAction,
+        expects: defaultExpected,
+        config: undefined,
+        message,
+      } satisfies ParseBooleanAction<unknown, undefined, () => string>);
+    });
 
-      expect(stringbool(options)).toStrictEqual({
-        kind: 'transformation',
-        type: 'stringbool',
-        reference: stringbool,
+    test('with config and undefined message', () => {
+      const action: ParseBooleanAction<unknown, typeof config, undefined> = {
+        ...baseAction,
         expects: '("yep" | "nope")',
-        options: {
-          truthy: ['yep'],
-          falsy: ['nope'],
-          case: 'insensitive',
-        },
-        async: false,
-        '~run': expect.any(Function),
-      } satisfies StringboolAction<unknown>);
+        config,
+        message: undefined,
+      };
+      expect(parseBoolean(config)).toStrictEqual(action);
+      expect(parseBoolean(config, undefined)).toStrictEqual(action);
     });
 
-    test('for custom options all together', () => {
-      const options: StringboolOptions = {
-        truthy: ['YEP'],
-        falsy: ['NOPE'],
-        case: 'sensitive',
-      };
+    test('with config and string message', () => {
+      expect(parseBoolean(config, 'message')).toStrictEqual({
+        ...baseAction,
+        expects: '("yep" | "nope")',
+        config,
+        message: 'message',
+      } satisfies ParseBooleanAction<unknown, typeof config, 'message'>);
+    });
 
-      expect(stringbool(options)).toStrictEqual({
-        kind: 'transformation',
-        type: 'stringbool',
-        reference: stringbool,
-        expects: '("YEP" | "NOPE")',
-        options,
-        async: false,
-        '~run': expect.any(Function),
-      } satisfies StringboolAction<unknown>);
+    test('with config and function message', () => {
+      const message = () => 'message';
+      expect(parseBoolean(config, message)).toStrictEqual({
+        ...baseAction,
+        expects: '("yep" | "nope")',
+        config,
+        message,
+      } satisfies ParseBooleanAction<unknown, typeof config, () => string>);
     });
   });
 
-  describe('should handle default options', () => {
-    const action = stringbool();
+  describe('should handle default config', () => {
+    const action = parseBoolean();
 
     describe('for truthy values', () => {
       test('for "true"', () => {
@@ -144,6 +144,20 @@ describe('stringbool', () => {
         expect(
           action['~run']({ typed: true, value: 'TRUE' }, {})
         ).toStrictEqual({
+          typed: true,
+          value: true,
+        });
+      });
+
+      test('for boolean true', () => {
+        expect(action['~run']({ typed: true, value: true }, {})).toStrictEqual({
+          typed: true,
+          value: true,
+        });
+      });
+
+      test('for number 1', () => {
+        expect(action['~run']({ typed: true, value: 1 }, {})).toStrictEqual({
           typed: true,
           value: true,
         });
@@ -207,12 +221,28 @@ describe('stringbool', () => {
           value: false,
         });
       });
+
+      test('for boolean false', () => {
+        expect(action['~run']({ typed: true, value: false }, {})).toStrictEqual(
+          {
+            typed: true,
+            value: false,
+          }
+        );
+      });
+
+      test('for number 0', () => {
+        expect(action['~run']({ typed: true, value: 0 }, {})).toStrictEqual({
+          typed: true,
+          value: false,
+        });
+      });
     });
   });
 
-  describe('should handle custom options with `truthy` only', () => {
+  describe('should handle custom config with `truthy` only', () => {
     describe('for lowercase `truthy` values', () => {
-      const action = stringbool({
+      const action = parseBoolean({
         truthy: ['yep'],
       });
 
@@ -245,7 +275,7 @@ describe('stringbool', () => {
     });
 
     describe('for uppercase `truthy` values', () => {
-      const action = stringbool({
+      const action = parseBoolean({
         truthy: ['YEP'],
       });
 
@@ -269,9 +299,9 @@ describe('stringbool', () => {
     });
   });
 
-  describe('should handle custom options with `falsy` only', () => {
+  describe('should handle custom config with `falsy` only', () => {
     describe('for lowercase `falsy` values', () => {
-      const action = stringbool({
+      const action = parseBoolean({
         falsy: ['nope'],
       });
 
@@ -304,7 +334,7 @@ describe('stringbool', () => {
     });
 
     describe('for uppercase `falsy` values', () => {
-      const action = stringbool({
+      const action = parseBoolean({
         falsy: ['NOPE'],
       });
 
@@ -328,79 +358,10 @@ describe('stringbool', () => {
     });
   });
 
-  describe('should handle custom options with `case` only', () => {
-    describe('for case sensitive', () => {
-      const action = stringbool({
-        case: 'sensitive',
-      });
-
-      test('for "true"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'true' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: true,
-        });
-      });
-
-      test('for "false"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'false' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: false,
-        });
-      });
-    });
-
-    describe('for case insensitive', () => {
-      const action = stringbool({
-        case: 'insensitive',
-      });
-
-      test('for "true"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'true' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: true,
-        });
-      });
-
-      test('for "TRUE"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'TRUE' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: true,
-        });
-      });
-
-      test('for "false"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'false' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: false,
-        });
-      });
-
-      test('for "FALSE"', () => {
-        expect(
-          action['~run']({ typed: true, value: 'FALSE' }, {})
-        ).toStrictEqual({
-          typed: true,
-          value: false,
-        });
-      });
-    });
-  });
-
-  describe('should handle custom options all together', () => {
-    const action = stringbool({
+  describe('should handle custom config all together', () => {
+    const action = parseBoolean({
       truthy: ['YEP'],
       falsy: ['NOPE'],
-      case: 'sensitive',
     });
 
     test('for "YEP"', () => {
@@ -418,16 +379,58 @@ describe('stringbool', () => {
     });
   });
 
+  describe('should handle custom non-string config', () => {
+    const action = parseBoolean({
+      truthy: [1, true],
+      falsy: [0, false, null],
+    });
+
+    test('for number 1', () => {
+      expect(action['~run']({ typed: true, value: 1 }, {})).toStrictEqual({
+        typed: true,
+        value: true,
+      });
+    });
+
+    test('for boolean true', () => {
+      expect(action['~run']({ typed: true, value: true }, {})).toStrictEqual({
+        typed: true,
+        value: true,
+      });
+    });
+
+    test('for number 0', () => {
+      expect(action['~run']({ typed: true, value: 0 }, {})).toStrictEqual({
+        typed: true,
+        value: false,
+      });
+    });
+
+    test('for boolean false', () => {
+      expect(action['~run']({ typed: true, value: false }, {})).toStrictEqual({
+        typed: true,
+        value: false,
+      });
+    });
+
+    test('for null', () => {
+      expect(action['~run']({ typed: true, value: null }, {})).toStrictEqual({
+        typed: true,
+        value: false,
+      });
+    });
+  });
+
   describe('should return dataset with issues', () => {
-    describe('for default options', () => {
-      const action = stringbool();
+    describe('for default config', () => {
+      const action = parseBoolean();
 
       const baseIssue: Omit<
-        StringboolIssue<string>,
+        ParseBooleanIssue<string>,
         'input' | 'received' | 'message'
       > = {
         kind: 'transformation',
-        type: 'stringbool',
+        type: 'parse_boolean',
         expected: defaultExpected,
         requirement: undefined,
         path: undefined,
@@ -448,7 +451,7 @@ describe('stringbool', () => {
               ...baseIssue,
               input: value,
               received: 'symbol',
-              message: `Invalid stringbool: Expected ${defaultExpected} but received symbol`,
+              message: `Invalid boolean: Expected ${defaultExpected} but received symbol`,
             },
           ],
         });
@@ -465,7 +468,7 @@ describe('stringbool', () => {
               ...baseIssue,
               input: value,
               received: '"something"',
-              message: `Invalid stringbool: Expected ${defaultExpected} but received "something"`,
+              message: `Invalid boolean: Expected ${defaultExpected} but received "something"`,
             },
           ],
         });
@@ -482,25 +485,26 @@ describe('stringbool', () => {
               ...baseIssue,
               input: value,
               received: '123',
-              message: `Invalid stringbool: Expected ${defaultExpected} but received 123`,
+              message: `Invalid boolean: Expected ${defaultExpected} but received 123`,
             },
           ],
         });
       });
     });
 
-    describe('for custom options truthy only', () => {
-      const action = stringbool({
+    describe('for custom config truthy only', () => {
+      const action = parseBoolean({
         truthy: ['yep'],
       });
 
       const baseIssue: Omit<
-        StringboolIssue<string>,
+        ParseBooleanIssue<string>,
         'input' | 'received' | 'message'
       > = {
         kind: 'transformation',
-        type: 'stringbool',
-        expected: '("yep" | "false" | "0" | "no" | "n" | "off" | "disabled")',
+        type: 'parse_boolean',
+        expected:
+          '("yep" | false | 0 | "false" | "0" | "no" | "n" | "off" | "disabled")',
         requirement: undefined,
         path: undefined,
         issues: undefined,
@@ -520,25 +524,26 @@ describe('stringbool', () => {
               ...baseIssue,
               input: value,
               received: '"enabled"',
-              message: `Invalid stringbool: Expected ${baseIssue.expected} but received "enabled"`,
+              message: `Invalid boolean: Expected ${baseIssue.expected} but received "enabled"`,
             },
           ],
         });
       });
     });
 
-    describe('for custom options falsy only', () => {
-      const action = stringbool({
+    describe('for custom config falsy only', () => {
+      const action = parseBoolean({
         falsy: ['nope'],
       });
 
       const baseIssue: Omit<
-        StringboolIssue<string>,
+        ParseBooleanIssue<string>,
         'input' | 'received' | 'message'
       > = {
         kind: 'transformation',
-        type: 'stringbool',
-        expected: '("true" | "1" | "yes" | "y" | "on" | "enabled" | "nope")',
+        type: 'parse_boolean',
+        expected:
+          '(true | 1 | "true" | "1" | "yes" | "y" | "on" | "enabled" | "nope")',
         requirement: undefined,
         path: undefined,
         issues: undefined,
@@ -558,26 +563,25 @@ describe('stringbool', () => {
               ...baseIssue,
               input: value,
               received: '"disabled"',
-              message: `Invalid stringbool: Expected ${baseIssue.expected} but received "disabled"`,
+              message: `Invalid boolean: Expected ${baseIssue.expected} but received "disabled"`,
             },
           ],
         });
       });
     });
 
-    describe('for custom options all together', () => {
-      const action = stringbool({
+    describe('for custom config all together', () => {
+      const action = parseBoolean({
         truthy: ['YEP'],
         falsy: ['NOPE'],
-        case: 'sensitive',
       });
 
       const baseIssue: Omit<
-        StringboolIssue<string>,
+        ParseBooleanIssue<string>,
         'input' | 'received' | 'message'
       > = {
         kind: 'transformation',
-        type: 'stringbool',
+        type: 'parse_boolean',
         expected: '("YEP" | "NOPE")',
         requirement: undefined,
         path: undefined,
@@ -587,8 +591,8 @@ describe('stringbool', () => {
         abortPipeEarly: undefined,
       };
 
-      test('for invalid lowercase input as string "yep"', () => {
-        const value = 'yep';
+      test('for invalid input as string "something"', () => {
+        const value = 'something';
 
         expect(action['~run']({ typed: true, value }, {})).toStrictEqual({
           typed: false,
@@ -597,25 +601,63 @@ describe('stringbool', () => {
             {
               ...baseIssue,
               input: value,
-              received: '"yep"',
-              message: `Invalid stringbool: Expected ${baseIssue.expected} but received "yep"`,
+              received: '"something"',
+              message: `Invalid boolean: Expected ${baseIssue.expected} but received "something"`,
+            },
+          ],
+        });
+      });
+    });
+
+    describe('with custom message', () => {
+      test('with string message', () => {
+        const action = parseBoolean(undefined, 'custom message');
+        const value = 'something';
+
+        expect(action['~run']({ typed: true, value }, {})).toStrictEqual({
+          typed: false,
+          value,
+          issues: [
+            {
+              kind: 'transformation',
+              type: 'parse_boolean',
+              expected: defaultExpected,
+              input: value,
+              received: '"something"',
+              message: 'custom message',
+              requirement: undefined,
+              path: undefined,
+              issues: undefined,
+              lang: undefined,
+              abortEarly: undefined,
+              abortPipeEarly: undefined,
             },
           ],
         });
       });
 
-      test('for invalid lowercase input as string "nope"', () => {
-        const value = 'nope';
+      test('with function message', () => {
+        const message = () => 'custom message';
+        const action = parseBoolean(undefined, message);
+        const value = 'something';
 
         expect(action['~run']({ typed: true, value }, {})).toStrictEqual({
           typed: false,
           value,
           issues: [
             {
-              ...baseIssue,
+              kind: 'transformation',
+              type: 'parse_boolean',
+              expected: defaultExpected,
               input: value,
-              received: '"nope"',
-              message: `Invalid stringbool: Expected ${baseIssue.expected} but received "nope"`,
+              received: '"something"',
+              message: 'custom message',
+              requirement: undefined,
+              path: undefined,
+              issues: undefined,
+              lang: undefined,
+              abortEarly: undefined,
+              abortPipeEarly: undefined,
             },
           ],
         });